refactor: random -> getRandomValue로 함수 이름 변경
